# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Task Trail API",
        "contact": {
            "name": "HiRaise",
            "url": "https://hiraise.net/",
            "email": "musaev.ae@hiraise.net"
        },
        "license": {
            "name": "MIT License",
            "url": "https://mit-license.org/"
        },
        "version": "1.0"
    },
    "paths": {
        "/v1/auth/check": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/v1/auth"
                ],
                "summary": "check user authentication",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "authentication required",
                        "schema": {
                            "$ref": "#/definitions/response.ErrAPI"
                        }
                    }
                }
            }
        },
        "/v1/auth/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/v1/auth"
                ],
                "summary": "login user",
                "parameters": [
                    {
                        "description": "user email and password",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.credentials"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "invalid request body",
                        "schema": {
                            "$ref": "#/definitions/response.ErrAPI"
                        }
                    },
                    "401": {
                        "description": "invalid credentials",
                        "schema": {
                            "$ref": "#/definitions/response.ErrAPI"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrAPI"
                        }
                    }
                }
            }
        },
        "/v1/auth/logout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/v1/auth"
                ],
                "summary": "logout user",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "authentication required",
                        "schema": {
                            "$ref": "#/definitions/response.ErrAPI"
                        }
                    }
                }
            }
        },
        "/v1/auth/password/change": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/v1/auth"
                ],
                "summary": "change user password",
                "parameters": [
                    {
                        "description": "old and new password",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.changePasswordReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "invalid request body",
                        "schema": {
                            "$ref": "#/definitions/response.ErrAPI"
                        }
                    }
                }
            }
        },
        "/v1/auth/password/forgot": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/v1/auth"
                ],
                "summary": "send reset password email",
                "parameters": [
                    {
                        "description": "user email",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.emailReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "invalid request body",
                        "schema": {
                            "$ref": "#/definitions/response.ErrAPI"
                        }
                    }
                }
            }
        },
        "/v1/auth/password/reset": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/v1/auth"
                ],
                "summary": "reset user password",
                "parameters": [
                    {
                        "description": "token and new password",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.resetPasswordReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "invalid request body",
                        "schema": {
                            "$ref": "#/definitions/response.ErrAPI"
                        }
                    }
                }
            }
        },
        "/v1/auth/refresh": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/v1/auth"
                ],
                "summary": "refresh tokens pair",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "refresh token is invalid",
                        "schema": {
                            "$ref": "#/definitions/response.ErrAPI"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrAPI"
                        }
                    }
                }
            }
        },
        "/v1/auth/register": {
            "post": {
                "description": "endpoint for register new user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/v1/auth"
                ],
                "summary": "register new user",
                "parameters": [
                    {
                        "description": "user email and password",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.credentials"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "invalid request body",
                        "schema": {
                            "$ref": "#/definitions/response.ErrAPI"
                        }
                    },
                    "409": {
                        "description": "user already exists",
                        "schema": {
                            "$ref": "#/definitions/response.ErrAPI"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrAPI"
                        }
                    }
                }
            }
        },
        "/v1/auth/resend-verification": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/v1/auth"
                ],
                "summary": "resend account verification email",
                "parameters": [
                    {
                        "description": "user email",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.emailReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "invalid request body",
                        "schema": {
                            "$ref": "#/definitions/response.ErrAPI"
                        }
                    }
                }
            }
        },
        "/v1/auth/verify": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/v1/auth"
                ],
                "summary": "verify user account",
                "parameters": [
                    {
                        "description": "token",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.verifyReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "token is invalid",
                        "schema": {
                            "$ref": "#/definitions/response.ErrAPI"
                        }
                    },
                    "404": {
                        "description": "token or user not found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrAPI"
                        }
                    }
                }
            }
        },
        "/v1/users/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "...",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/v1/users"
                ],
                "summary": "return current user",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "...",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/v1/users"
                ],
                "summary": "update current user",
                "parameters": [
                    {
                        "description": "user data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "invalid request body",
                        "schema": {
                            "$ref": "#/definitions/response.ErrAPI"
                        }
                    },
                    "401": {
                        "description": "authentication required",
                        "schema": {
                            "$ref": "#/definitions/response.ErrAPI"
                        }
                    },
                    "404": {
                        "description": "user not found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrAPI"
                        }
                    }
                }
            }
        },
        "/v1/users/me/avatar": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "...",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/v1/users"
                ],
                "summary": "upload new avatar",
                "parameters": [
                    {
                        "type": "file",
                        "description": "new file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "authentication required",
                        "schema": {
                            "$ref": "#/definitions/response.ErrAPI"
                        }
                    }
                }
            }
        },
        "/v1/users/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "...",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/v1/users"
                ],
                "summary": "return user by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "user id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "definitions": {
        "request.UpdateReq": {
            "type": "object",
            "properties": {
                "username": {
                    "type": "string",
                    "maxLength": 100
                }
            }
        },
        "request.changePasswordReq": {
            "type": "object",
            "required": [
                "newPassword",
                "oldPassword"
            ],
            "properties": {
                "newPassword": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 8
                },
                "oldPassword": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 8
                }
            }
        },
        "request.credentials": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 8
                }
            }
        },
        "request.emailReq": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
        "request.resetPasswordReq": {
            "type": "object",
            "required": [
                "password",
                "token"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 8
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "request.verifyReq": {
            "type": "object",
            "required": [
                "token"
            ],
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
        "response.ErrAPI": {
            "type": "object",
            "properties": {
                "metadata": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "msg": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "at",
            "in": "cookie"
        }
    }
}